agent/requirements.txt
certifi==2019.11.28
chardet==3.0.4
grpcio==1.27.2
grpcio-tools==1.27.2
idna==2.9
protobuf==3.11.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
